requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
assemblyai
google-generativeai
jinja2
python-multipart